refactor tests folder
