+++ functions/tests/functions/misc/health.test.ts
@@ -1,8 +1,8 @@
 import { program } from 'commander';
-import { Context } from './models/context';
-import { FirebaseClient } from './service/firebase-client';
+import { Context } from '../../models/context';
+import { FirebaseClient } from '../../service/firebase-client';
 import { z } from 'zod';
-import { Environments } from './models/environments';
+import { Environments } from '../../models/environments';
 import assert from 'assert';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
