+++ functions/tests/functions/user/create-fcm-token.test.ts
@@ -1,12 +1,12 @@
 import { program } from 'commander';
-import { Context } from './models/context';
-import { FirebaseClient } from './service/firebase-client';
+import { Context } from '../../models/context';
+import { FirebaseClient } from '../../service/firebase-client';
 import { z } from 'zod';
-import { NonEmptyString } from '../src/models/non-empty-string';
-import { Environments } from './models/environments';
-import { Firebase } from '../src/services/firebase';
-import { Collection } from '../src/models/collection-name';
-import { Utils } from '../src/utils/utils';
+import { NonEmptyString } from '../../../src/models/non-empty-string';
+import { Environments } from '../../models/environments';
+import { Firebase } from '../../../src/services/firebase';
+import { Collection } from '../../../src/models/collection-name';
+import { Utils } from '../../../src/utils/utils';
 import assert from 'assert';
 
 export const createFCMToken = async (ctx: Context, reqData: any) => {
