+++ functions/tests/functions/user/create-test-user.test.ts
@@ -1,11 +1,11 @@
 import { program } from 'commander';
-import { Context } from './models/context';
-import { FirebaseClient } from './service/firebase-client';
+import { Context } from '../../models/context';
+import { FirebaseClient } from '../../service/firebase-client';
 import { z } from 'zod';
-import { NonEmptyString } from '../src/models/non-empty-string';
-import { Environments } from './models/environments';
+import { NonEmptyString } from '../../../src/models/non-empty-string';
+import { Environments } from '../../models/environments';
 import assert from 'assert';
-import { Utils } from '../src/utils/utils';
+import { Utils } from '../../../src/utils/utils';
 
 export const createTestUser = async (ctx: Context, reqData: any) => {
 	const res = await ctx.firebase.functions('createTestUser')(reqData);
